requirements.txt
docopt==0.6.2
stomp.py==8.1.0
websocket-client==1.5.1
